Move blog app serializers to api app
