+++ api/urls.py
@@ -1,10 +1,20 @@
 from django.conf.urls import url
 
 from .views import (
+    # account
     ChangeUserPasswordView,
-    MainUserFieldsView, UsersView,
-    UserLoginAPI, UserRegistrationAPI, UserForgetPasswordAPI,
+    MainUserFieldsView,
+    UsersView,
+    UserLoginAPI,
+    UserRegistrationAPI,
+    UserForgetPasswordAPI,
     UserShortInfoAPI,
+    # blog
+    ArticlesView,
+    ArticlesAllAPI,
+    ArticleCreateAPI,
+    ArticleRemoveAPI,
+    ArticlesSearchAPI
 )
 
 urlpatterns = [
@@ -17,4 +27,11 @@ urlpatterns = [
     url(r'^login-api/$', UserLoginAPI.as_view()),
     url(r'^user-registration-api/$', UserRegistrationAPI.as_view()),
     url(r'^user-forget-password-api/$', UserForgetPasswordAPI.as_view()),
+    # blog API
+    url(r'^articles-rest/$', ArticlesView.as_view()),
+    url(r'^articles-rest/(?P<pk>\d+)/$', ArticlesView.as_view()),
+    url(r'^articles-all-api/$', ArticlesAllAPI.as_view()),
+    url(r'^articles-create-api/$', ArticleCreateAPI.as_view()),
+    url(r'^articles-remove-api/$', ArticleRemoveAPI.as_view()),
+    url(r'^articles-search-api/$', ArticlesSearchAPI.as_view()),
 ]
